+++ mir-extractor/src/rules/input.rs
@@ -523,7 +523,7 @@ impl Rule for InfiniteIteratorRule {
                     function: function.name.clone(),
                     function_signature: function.signature.clone(),
                     evidence,
-                    span: None,
+                    span: function.span.clone(),
                     ..Default::default()
                 });
             }